move extensions to detecated folder
